refactor: Split decode and instruction to isa/ and refactor some define

+++ crates/riscv-core/src/error.rs
@@ -17,11 +17,11 @@ pub enum RiscVError {
     #[error("PC: PC value `{0}` not misaligned to byte")]
     InstructionAddressMisaligned(u32),
 
-    #[error("OpCode: Not implemented opcode: 0x{0:x}")]
-    NotImplementedOpCode(u32),
+    #[error("OpCode: Not implemented opcode: 0x{0:02x}")]
+    NotImplementedOpCode(u8),
 
-    #[error("OpCode: Not implemented func 0x{0:x} from: 0x{1:x}")]
-    NotImplementedFunc(u32, u8),
+    #[error("OpCode: Not implemented funct 0x{0:x} from: 0x{1:x}")]
+    NotImplementedFunc(u8, u8),
 
     #[error("System Call: Not implemented system call: {0}")]
     NotImplementedSysCall(u32),
